requirements.txt
altair==6.0.0
aptdaemon==2.0.2
attrs==25.3.0
autocommand==2.2.2
babel==2.17.0
bcc==0.31.0
bcrypt==4.2.0
blinker==1.9.0
Brlapi==0.8.6
cachetools==6.2.6
certifi==2025.1.31
chardet==5.2.0
click==8.1.8
command-not-found==0.3
configobj==5.0.9
contourpy==1.3.3
cryptography==43.0.0
cupshelpers==1.0
cycler==0.12.1
dasbus==1.7
dbus-python==1.4.0
decorator==5.2.1
defer==1.0.6
distro==1.9.0
distro-info==1.14
fonttools==4.61.1
gitdb==4.0.12
GitPython==3.1.46
httplib2==0.22.0
idna==3.10
inflect==7.3.1
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
jsonpatch==1.32
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.12.1
kiwisolver==1.4.9
language-selector==0.1
launchpadlib==2.1.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
linkify-it-py==2.0.3
louis==3.33.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
mdurl==0.1.2
more-itertools==10.7.0
narwhals==2.16.0
netaddr==1.3.0
numpy==2.4.2
oauthlib==3.2.2
olefile==0.47
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pexpect==4.9.0
pillow==11.3.0
plotly==6.5.2
prometheus_client==0.21.1
protobuf==6.33.5
psutil==7.0.0
ptyprocess==0.7.0
pyarrow==23.0.0
pycairo==1.27.0
pycups==2.0.4
pydeck==0.9.1
Pygments==2.18.0
PyGObject==3.50.0
PyJWT==2.10.1
pyparsing==3.1.2
pyserial==3.5
python-apt==3.0.0+ubuntu1.1
python-dateutil==2.9.0
python-debian==1.0.1+ubuntu1
pytz==2025.2
pyxdg==0.28
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
setproctitle==1.3.6
smmap==5.0.2
streamlit==1.54.0
systemd-python==235
tenacity==9.1.4
toml==0.10.2
tornado==6.5.4
typeguard==4.4.2
typing_extensions==4.13.2
tzdata==2025.3
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
uc-micro-py==1.0.3
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.3.0
wadllib==2.0.0
watchdog==6.0.0
wheel==0.46.1
xdg==5
xkit==0.0.0
